+++ bitmap_test.c
@@ -100,8 +100,8 @@ MU_TEST(bitmap_test_tiling)
 	long	test1_ticks;
 	long	test2_ticks;
 	long	test3_ticks;
-	int		i;
-	int		times_to_run = 100;
+	int16_t	i;
+	int16_t	times_to_run = 100;
 	
 	Screen* the_screen = Sys_GetScreen(global_system, ID_CHANNEL_B);
 	Theme*	the_theme = Sys_GetTheme(global_system);
